Release/201612142113 pull from wes bos (#1)
* Fix typos and extra code

* add the instant function call

* changes

* Fixes titles

* 💩

* adding women inventors

This is a random selection of female inventors for the dataset

* female inventors + change in const name

A random selection of female inventors for the dataset along with proposed change in const name.

* Unused CSS :smile:

* fixed hours

* typo: inventory > inventors'

* Fix typo in readme.md

* Removed the extra code from the start file

* 7. sort Exercise make it so it sorts on last name 

In your video you still sort on the first name of the array but you called it `aLast` and `bLast` but they where in the first position of the array like this: `const [aLast, aFirst]` when changed to `const [aFirst, aLast]` it will filter on the second part of the array. There for it will sort on the last name. 

To bad there is no way to pull request your video.

* Update readme.md

As requested by Wes Bos.

* Massages Array Cardio challenges 💪

* Typo fix

* 24 - Sticky Nav: reset index and style START

Reset the contents of index-START and style-START to reflect the
contents before running through the video.

* Update index.html

spelling.

* copy-pasta typo

* Only START and FINISHED files are necessary.